requirement.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
django-leaflet==0.28.3
djangorestframework==3.14.0
psycopg2-binary==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
